%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  findCfitBinary.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{findCfitBinary}
\alias{findCfitBinary}


\title{Located the cfit binary}

\description{
  Located the cfit binary and sets option 'cfit'.
}

\usage{
findCfitBinary(pkgname="sfit", quote=TRUE, force=FALSE, ...)
}

\arguments{
  \item{pkgname}{A \code{\link[base]{character}} string specifying the name of the
    package to be searched.}
  \item{quote}{If \code{\link[base:logical]{FALSE}}, the returned string is not quoted.}
  \item{force}{If \code{\link[base:logical]{FALSE}} and a previously located binary exists,
    then it is used.}
  \item{...}{Not used.}
}

\value{
  Returns (invisibly) a \code{\link[base]{character}} string.
}


\keyword{internal}
