%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.NonDocumentedObjects.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{Non-documented objects}
\alias{Non-documented objects}
\title{Non-documented objects}


% Plot methods:
\alias{getEdges}
\alias{getEdges.cfit}
\alias{points.cfit}
\alias{points3d}
\alias{points3d.cfit}
\alias{lines.cfit}
\alias{lines3d}
\alias{lines3d.cfit}
\alias{radials}
\alias{radials.cfit}
\alias{drawApex}
\alias{drawApex.cfit}

% Misc methods:
\alias{cfitOptions}
\alias{cfitTests}

\description{
  This page contains aliases for all "non-documented" objects that
  \code{R CMD check} detects in this package.

  Most of the are to be documented.
}

\author{Henrik Bengtsson}


\keyword{documentation}
\keyword{internal}
