%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.package.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{sfit-package}
\alias{sfit-package}
\alias{sfit}
\docType{package}

\title{Package sfit}


\description{
  Multidimensional simplex fitting.
}

\section{Requirements}{
  This package requires the package \pkg{R.oo}.
}

\section{Installation and updates}{

  To install this package, see \url{http://www.braju.com/R/}.
}

\section{To get started}{
  To get started, see:
  \enumerate{
    \item \code{\link{cfit.matrix}}() - To fit a K-dimensional simplex in an
       N-dimensional space.
  }
}

\section{How to cite this package}{
  Please site [1] and [2] below.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\references{
 [1] P. Wirapati, & T. Speed, \emph{Fitting polyhedrial cones and
    simplices to multivariate data points}, Walter and Eliza Hall Institute
    of Medical Research, December 30, 2001.\cr
 [2] P. Wirapati and T. Speed, \emph{An algorithm to fit a simplex
    to a set of multidimensional points}, Walter and Eliza Hall Institute
    of Medical Research, January 15, 2002.\cr
}

\section{License}{
  The releases of this package is licensed under
  LGPL version 2.1 or newer.

  The development code of the packages is under a private licence
  (where applicable) and patches sent to the author fall under the
  latter license, but will be, if incorporated, released under the
  "release" license above.
}
\keyword{package}
